refactor into profile
